refactor: useRecordForm에서 useContent 분리
