+++ src/app/content/hooks/useRecordForm.ts
@@ -1,6 +1,5 @@
 import { USERNAMES } from '@/constants';
 import api from '@/api';
-import { useContent } from './useContent';
 import { useMutation } from '@tanstack/react-query';
 import { useState } from 'react';
 import { useTimer } from './useTimer';
@@ -8,30 +7,17 @@ import { useToast } from '@/components/ui/use-toast';
 
 export const INPUT_MAX_LENGTH = 100;
 
-// TODO: refactor
 export const useRecordForm = () => {
   const [input, setInput] = useState<string>('');
   const [username, setUsername] = useState<(typeof USERNAMES)[number]>(
     USERNAMES[0],
   );
 
-  const { content, fetchNewContent } = useContent();
-
-  const {
-    setIsActive,
-    remainingSeconds,
-    resetRemainingSeconds,
-    scheduleSetInactive,
-    isTimerEnd,
-  } = useTimer();
-
-  const { toast } = useToast();
-
   const { mutateAsync: createRecord } = useMutation({
-    mutationFn: async () => {
+    mutationFn: async (contentId: number) => {
       try {
         await api.post('/records/create', {
-          content_id: content?.id ?? -1,
+          content_id: contentId,
           username,
           text: input,
         });
@@ -40,6 +26,18 @@ export const useRecordForm = () => {
       }
     },
   });
+
+  const {
+    setIsActive,
+    remainingSeconds,
+    isTimerEnd,
+    resetRemainingSeconds,
+    scheduleSetInactive,
+  } = useTimer();
+
+  const { toast } = useToast();
+
+  const exceedsMaxLength = input.length > INPUT_MAX_LENGTH;
 
   const handleInputFocus = () => {
     setIsActive(true);
@@ -55,46 +53,42 @@ export const useRecordForm = () => {
     setIsActive(false);
   };
 
-  const resetForm = () => {
-    resetRemainingSeconds();
-    setIsActive(false);
-    setInput('');
-  };
-
-  const getNewContent = async () => {
-    await fetchNewContent();
-    resetForm();
-  };
-
-  const handleSubmit = async () => {
+  const submitForm = async (contentId: number) => {
     if (!input.length) {
-      toast({
+      return toast({
         description: '내용을 입력해 주세요',
       });
-    } else if (input.length > INPUT_MAX_LENGTH) {
-      toast({
-        description: '최대 100자를 입력해 주세요',
-      });
-    } else {
-      await createRecord();
-      toast({
-        description: '기록 완료!',
-      });
-      await getNewContent();
     }
+
+    if (input.length > INPUT_MAX_LENGTH) {
+      return toast({
+        description: '최대 100자를 입력해 주요',
+      });
+    }
+
+    await createRecord(contentId);
+    toast({
+      description: '기록 완료!',
+    });
+  };
+
+  const resetForm = () => {
+    setInput('');
+    resetRemainingSeconds();
+    setIsActive(false);
   };
 
   return {
-    content,
-    getNewContent,
+    input,
+    username,
+    setUsername,
     remainingSeconds,
     isTimerEnd,
-    input,
+    exceedsMaxLength,
     handleInputFocus,
     handleInputChange,
     handleInputBlur,
-    handleSubmit,
-    username,
-    setUsername,
+    submitForm,
+    resetForm,
   };
 };
